Move interfaces and enums into types folder
