+++ src/controllers/asset_pools/getAll.action.ts
@@ -2,6 +2,6 @@ import { Request, Response } from 'express';
 import AssetPoolService from '@/services/AssetPoolService';
 
 export const getAssetPools = async (req: Request, res: Response) => {
-    const result = await AssetPoolService.getAll(req.user.sub);
+    const result = await AssetPoolService.getAllBySub(req.user.sub);
     res.json(result);
 };
