--- src/enums/index.ts
@@ -1,5 +1,0 @@
-import { DepositState } from './DepositState';
-import { WithdrawalState } from './WithdrawalState';
-import { WithdrawalType } from './WithdrawalType';
-
-export { WithdrawalType, WithdrawalState, DepositState };
