+++ src/swagger.json
@@ -1,7 +1,7 @@
 {
   "info": {
     "title": "THX API Specification",
-    "version": "1.2.0"
+    "version": "1.2.5"
   },
   "basePath": "/v1",
   "securityDefinitions": {
