+++ src/types/definitions/qrcode.d.ts
