+++ src/types/interfaces/IPromoCodeData.ts
