+++ src/types/interfaces/IPromoCodeResponse.ts
