+++ src/types/interfaces/IWithdrawPoll.ts
